pyproject.toml
[project]
name = "simple-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docx2txt>=0.9",
    "faiss-cpu>=1.11.0.post1",
    "fastapi>=0.116.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-openai>=0.3.28",
    "loguru>=0.7.3",
    "openai>=1.98.0",
    "pydantic>=2.11.7",
    "pypdf>=5.9.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "tiktoken>=0.9.0",
    "uvicorn>=0.35.0",
]

[tool.uv]
index-url = "https://pypi.tuna.tsinghua.edu.cn/simple"
